requirements.txt
asn1crypto==1.2.0
astroid==2.3.2
cffi==1.13.1
cryptography==2.8
Django==2.2.6
django-extensions==2.2.5
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.3
python-decouple==3.1
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
